Changed commas to hyphens on line 3
Changing the commas to hyphens better illustrates your train of thought: the phrase "just seconds a day" is an interjection and not part of the rest of the phrase. I added the words "in order to" to better help with the flow of your sentence. I feel this commit improves the Readability of your readme file, and therefore improves the credibility of the project.